ref: split up listItemEditRequestValidation Squashed commits: [0afd02b5] list item [80f91471] remove test [dca20761] remove test [618b63e0] add back smoke flag [9032770b] ref: split up listItemEditRequestValidation
